frontend/requirements.txt
streamlit>=1.28.1
requests==2.31.0
pillow>=9.0.0
python-dotenv>=0.19.0
loguru>=0.7.3
aiohttp>=3.9.1

